Refactored ServiceFacadeTest into multiple tests
